Added parseFunctionDefinitions
